Move Surface_Marker and related types into surface_marker.py
